--- src/app/problem2/ProblemEditorContext.js
@@ -1,8 +1,0 @@
-import {createContext, useContext} from "react";
-
-export const ProblemEditorContext = createContext(undefined);
-
-export const useProblemEditorContext = () => {
-    return useContext(ProblemEditorContext);
-};
-
