--- src/app/problem2/commons/NewMarkdownEditorContext.js
@@ -1,7 +1,0 @@
-import {createContext, useContext} from "react";
-
-export const EditorContext = createContext(null);
-
-export const useEditorContext = () => {
-    return useContext(EditorContext);
-};
