--- src/app/problem2/left/leftbar.scss
@@ -1,8 +1,0 @@
-.problem-editor-leftbar {
-    font-family: Poppins, sans-serif;
-
-    a {
-        font-size: 17px;
-        color: rgba(164,164,164,1);
-    }
-}
