--- src/app/problem2/right/Actions.js
@@ -1,10 +1,0 @@
-import Block from "./Block";
-
-
-const Actions = () => {
-    return <>
-        <Block title="Actions" id="problem-editor-actions"/>
-    </>;
-};
-
-export default Actions;
