--- src/app/problem2/right/Divider.js
@@ -1,5 +1,0 @@
-
-
-export const Divider = () => {
-    return <hr className="padding-hr"/>;
-};
